backend/pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".venv"]
force-exclude = true

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.bandit]
exclude_dirs = [".venv"]
targets = ["."]

[project]
name = "backend"
version = "0.1.0"
description = "GraphQL API for wishlist and calendar"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.45",
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-auth>=2.45.0",
    "gunicorn>=23.0.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "strawberry-graphql>=0.287.2",
    "strawberry-graphql-django>=0.70.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.9",
]
